Product reviews

+++ FluxStore.Domain/Entities/ProductRating.cs
@@ -3,7 +3,7 @@
 	public class ProductRating
 	{
         public Guid Id { get; set; }
-        public int Rating { get; set; } // From 1 to 5
+        public double Rating { get; set; } // From 1 to 5
         public Guid ProductId { get; set; }
 
         public Product Product { get; set; } = default!;
